refactor(models): consolidate response handling
