+++ internal/models/doubao.go
@@ -6,7 +6,6 @@ import (
 	"net/http"
 
 	"github.com/Coien-rr/CommitWhisper/internal/comm"
-	selfErr "github.com/Coien-rr/CommitWhisper/pkg/errors"
 )
 
 const ARK_BASE_URL = "https://ark.cn-beijing.volces.com/api/v3"
@@ -59,14 +58,6 @@ type SessionError struct {
 		Param   string `json:"param"`
 		Type    string `json:"type"`
 	} `json:"error"`
-}
-
-type ResponseBody struct {
-	Choices []Choices `json:"choices"`
-}
-
-type Choices struct {
-	Message Message `json:"message"`
 }
 
 type errResponseBody struct {
@@ -208,38 +199,5 @@ func (m *DoubaoModel) GenerateCommitMessage(diffInfo string) (string, error) {
 		)
 	}
 
-	if statusCode == http.StatusOK {
-		var response ResponseBody
-		if err := json.Unmarshal(resp, &response); err != nil {
-			return "", fmt.Errorf(
-				"ERROR(generateCommitMessage): failed to parse response JSON: %v",
-				err,
-			)
-		}
-		return response.Choices[0].Message.Content, nil
-	} else {
-		var response errResponseBody
-		if err := json.Unmarshal(resp, &response); err != nil {
-			return "", fmt.Errorf(
-				"ERROR(generateCommitMessage): failed to parse response JSON: %v",
-				err,
-			)
-		}
-		switch statusCode {
-		case http.StatusUnauthorized:
-			// TODO: key Invalid  error
-			return "", selfErr.NewInvalidKeyError(response.ErrorMsg.Message)
-
-		case http.StatusNotFound:
-			// TODO: model not found error
-			return "", selfErr.NewNotFoundError(response.ErrorMsg.Message)
-
-		case http.StatusTooManyRequests:
-			// TODO: rate error or bill error
-			return "", selfErr.NewTooManyReqError(response.ErrorMsg.Message)
-
-		default:
-			return "", nil
-		}
-	}
+	return handleChatRespFromLLMs(resp, statusCode)
 }
